fix: update Snowflake integration tests and improve session handling
